refactor(deeplink): minor reformatting of the payment request
re #1325
