+++ src/actions/navigation/index.ts
@@ -9,12 +9,7 @@ import {
   NavigationNavigateAction,
 } from 'react-navigation'
 import { routeList } from 'src/routeList'
-import { JolocomLib } from 'jolocom-lib'
-import { interactionHandlers } from 'src/lib/storage/interactionTokens'
-import { AppError, ErrorCode } from 'src/lib/errors'
 import { ThunkAction } from 'src/store'
-import { keyIdToDid } from 'jolocom-lib/js/utils/helper'
-import { generateIdentitySummary } from '../sso/utils'
 
 const deferredNavActions: NavigationAction[] = []
 let dispatchNavigationAction = (action: NavigationAction) => {
@@ -89,39 +84,3 @@ export const navigatorReset = (
 export const navigatorResetHome = (): ThunkAction => dispatch =>
   dispatch(navigatorReset({ routeName: routeList.Home }))
 
-/**
- * The function that parses a deep link to get the route name and params
- * It then matches the route name and dispatches a corresponding action
- * @param jwt
- */
-export const handleDeepLink = (jwt: string): ThunkAction => async (
-  dispatch,
-  getState,
-  backendMiddleware,
-) => {
-  // TODO Fix
-  // The identityWallet is initialised before the deep link is handled. If it
-  // is not initialized, then we may not even have an identity.
-  if (!backendMiddleware.identityWallet) {
-    return dispatch(
-      navigate({
-        routeName: routeList.Landing,
-      }),
-    )
-  }
-
-  const interactionToken = JolocomLib.parse.interactionToken.fromJWT(jwt)
-  const issuer = await backendMiddleware.registry.resolve(
-    keyIdToDid(interactionToken.issuer),
-  )
-  const handler = interactionHandlers[interactionToken.interactionType]
-
-  if (!handler) {
-    throw new AppError(
-      ErrorCode.ParseJWTFailed,
-      new Error('Could not handle interaction token'),
-    )
-  }
-
-  return handler(interactionToken, generateIdentitySummary(issuer))
-}
