+++ src/lib/storage/interactionTokens.ts
@@ -7,7 +7,7 @@ import { Authentication } from 'jolocom-lib/js/interactionTokens/authentication'
 import { CredentialOfferRequest } from 'jolocom-lib/js/interactionTokens/credentialOfferRequest'
 import { CredentialRequest } from 'jolocom-lib/js/interactionTokens/credentialRequest'
 import { PaymentRequest } from 'jolocom-lib/js/interactionTokens/paymentRequest'
-import { consumePaymentRequest } from '../../actions/sso/paymentRequest'
+import { formatPaymentRequest } from '../../actions/sso/paymentRequest'
 import { IdentitySummary } from '../../actions/sso/types'
 /**
  * @param Metadata should not need to be passed to credential receive because it comes from cred Offer
@@ -34,5 +34,5 @@ export const interactionHandlers = {
   [InteractionType.PaymentRequest]: <T extends JSONWebToken<PaymentRequest>>(
     interactionToken: T,
     requesterSummary: IdentitySummary,
-  ) => consumePaymentRequest(interactionToken, requesterSummary),
+  ) => formatPaymentRequest(interactionToken, requesterSummary),
 }
